Improve augmentation in ckeditor5-restricted-editing.
