+++ packages/ckeditor5-restricted-editing/src/restrictededitingconfig.ts
@@ -62,16 +62,3 @@ export interface RestrictedEditingConfig {
 	 */
 	allowedAttributes: Array<string>;
 }
-
-declare module '@ckeditor/ckeditor5-core' {
-	interface EditorConfig {
-
-		/**
-		 * The configuration of the restricted editing mode feature. Introduced by the
-		 * {@link module:restricted-editing/restrictededitingmode~RestrictedEditingMode} feature.
-		 *
-		 * Read more in {@link module:restricted-editing/restrictededitingmode~RestrictedEditingModeConfig}.
-		 */
-		restrictedEditing?: RestrictedEditingConfig;
-	}
-}
