+++ packages/ckeditor5-restricted-editing/src/standardeditingmode.ts
@@ -35,8 +35,3 @@ export default class StandardEditingMode extends Plugin {
 	}
 }
 
-declare module '@ckeditor/ckeditor5-core' {
-	interface PluginsMap {
-		[ StandardEditingMode.pluginName ]: StandardEditingMode;
-	}
-}
