+++ packages/ckeditor5-restricted-editing/src/standardeditingmodeediting.ts
@@ -61,9 +61,3 @@ export default class StandardEditingModeEditing extends Plugin {
 		} );
 	}
 }
-
-declare module '@ckeditor/ckeditor5-core' {
-	interface PluginsMap {
-		[ StandardEditingModeEditing.pluginName ]: StandardEditingModeEditing;
-	}
-}
